+++ lib/jekyll/client_search/search_index_page.rb
@@ -1,11 +1,11 @@
 # frozen_string_literal: true
 
 module Jekyll
-  module ElasticlunrSearch
+  module ClientSearch
     class SearchIndexPage < Jekyll::PageWithoutAFile
       def initialize(site, output, documents)
         super(site, site.source, File.dirname(output), File.basename(output))
-        self.data = {}
+        self.data = { "layout" => nil, "sitemap" => false }
         self.content = JSON.generate(documents)
       end
     end