+++ lib/jekyll/elasticlunr_search.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-require "json"
-require "jekyll"
-require_relative "elasticlunr_search/version"
-require_relative "elasticlunr_search/configuration"
-require_relative "elasticlunr_search/document_builder"
-require_relative "elasticlunr_search/search_index_page"
-require_relative "elasticlunr_search/generator"