+++ spec/spec_helper.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require "bundler/setup"
-require "jekyll-elasticlunr-search"
+require "jekyll-client-search"
 
 RSpec.configure do |config|
   config.example_status_persistence_file_path = ".rspec_status"